refactor code for readability
