+++ packages/core/src/RenderingEngine/Viewport.ts
@@ -955,11 +955,12 @@ class Viewport implements IViewport {
 
     // update clippingPlanes if volume viewports
     const actorEntry = this.getDefaultActor();
-    if (actorEntry?.actor?.isA('vtkVolume')) {
-      this.updateClippingPlanesForActors(updatedCamera);
-    }
 
-    if (actorEntry?.actor?.isA('vtkImageSlice')) {
+    const isImageSlice = actorEntry?.actor?.isA('vtkImageSlice');
+
+    if (!isImageSlice) {
+      this.updateClippingPlanesForActors(updatedCamera);
+    } else {
       const renderer = this.getRenderer();
       renderer.resetCameraClippingRange();
     }
@@ -1007,7 +1008,7 @@ class Viewport implements IViewport {
       // we assume that the first two clipping plane of the mapper are always
       // the 'camera' clipping. Update clipping planes only if the actor is
       // a vtkVolume
-      if (!actorEntry.actor || !isImageActor(actorEntry.actor)) {
+      if (!actorEntry.actor) {
         return;
       }
